+++ src/engine/r_things.h
@@ -46,6 +46,6 @@ void R_RenderPlayerSprites(player_t* player);
 void R_DrawThingBBox(void);
 
 extern spritedef_t* spriteinfo;
-extern int            numsprites;
+extern intptr_t            numsprites;
 
 #endif
